+++ Changelog.md
@@ -2,6 +2,10 @@
 
 All significant changes are listed in this file.
 
+## [1.2.4] - 2024-8-17
+- Add actress scraper for AsianScreens.com
+- Improve scraper speed
+
 ## [1.2.3] - 2024-8-8
 - Add option to display Japanese name order
 - Fix issues with move / rename error handling